fix(apps/client): HASHI-80 하시픽 코드리뷰 반영
