+++ apps/client/src/features/restaurantList/constants/hashiPick.ts
@@ -1,0 +1,7 @@
+import type { FilterOption } from '../types'
+
+export const HASHI_PICK_SORT_OPTIONS: FilterOption[] = [
+  { label: '기본순', value: 'default' },
+  { label: '인기순', value: 'popular' },
+  { label: '별점순', value: 'rating' },
+]
