+++ apps/client/src/features/restaurantList/constants/mockRestaurants.ts
@@ -1,0 +1,279 @@
+import type { Restaurant } from '../types'
+
+export const MOCK_RESTAURANTS: Restaurant[] = [
+  {
+    id: '1',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '2',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '3',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '4',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '5',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '6',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '7',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '8',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '9',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '10',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '11',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '12',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '13',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '14',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '15',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '16',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '17',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '18',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '19',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '20',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '21',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '22',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '23',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '24',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+  {
+    id: '25',
+    name: '히마와리 스시 신도심점',
+    rating: 4.8,
+    region: '도쿄',
+    category: '초밥',
+    images: [],
+    description:
+      '식당 소개를 여기 간단하게 한 줄 적어주세요. 식당 소개를 여기 간단하게 한 줄 적어주세요.',
+    hashtags: ['#해시태그', '#해시태그', '#해시태그'],
+  },
+]
