+++ apps/client/src/features/restaurantList/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useInfiniteRestaurantList } from './useInfiniteRestaurantList'
