+++ apps/client/src/features/restaurantList/index.ts
@@ -1,3 +1,11 @@
-export { RestaurantListPage } from './RestaurantListPage'
-export { HASHI_PICK_SORT_OPTIONS, MOCK_RESTAURANTS } from './constants'
+export {
+  CATEGORY_OPTIONS,
+  DEFAULT_CATEGORY_OPTION,
+  HASHI_PICK_SORT_OPTIONS,
+  MOCK_RESTAURANTS,
+  RESTAURANT_LIST_PAGE_SIZE,
+} from './constants'
+export { useInfiniteRestaurantList } from './hooks'
+export { RestaurantCard } from './RestaurantCard'
+export { RestaurantFilterBar } from './RestaurantFilterBar'
 export type { FilterOption, Restaurant } from './types'
